# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Roil Finance API
  description: |
    Backend API for Roil Finance — an institutional-grade portfolio rebalancing
    platform built on Canton Network with Daml smart contracts.

    Supports portfolio management, DCA scheduling, reward tracking, market data,
    token transfers, yield compounding, and platform administration.
  version: 1.0.0
  contact:
    name: Roil Finance
    url: https://roil.app
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.roil.app
    description: Production

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing `actAs` array of Canton party identifiers.
        Algorithms supported: HS256 (dev), RS256, ES256.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Something went wrong"
...
    Portfolio:
      type: object
      properties:
        contractId:
          type: string
        user:
          type: string
        platform:
          type: string
        targets:
          type: array
          items:
            $ref: '#/components/schemas/TargetAllocation'
        holdings:
          type: array
          items:
            type: object
        triggerMode:
          $ref: '#/components/schemas/TriggerMode'
        isActive:
          type: boolean
...
paths:
...
  /api/portfolio/from-template:
    post:
      tags: [Portfolio]
      summary: Create portfolio from a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, templateId]
              properties:
                user:
                  type: string
                templateId:
                  type: string
      responses:
        '201':
          description: Portfolio created from template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/portfolio/{party}:
    get:
      tags: [Portfolio]
      summary: Get all portfolios for a party
      parameters:
        - name: party
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of portfolios
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Portfolio'
        '403':
          description: Not authorized for this party
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio:
    post:
      tags: [Portfolio]
      summary: Create a new portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, targets, triggerMode]
              properties:
                user:
                  type: string
                targets:
                  type: array
                  items:
                    $ref: '#/components/schemas/TargetAllocation'
                  minItems: 2
                triggerMode:
                  $ref: '#/components/schemas/TriggerMode'
      responses:
        '201':
          description: Portfolio created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error or targets do not sum to 100%
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/targets:
    put:
      tags: [Portfolio]
      summary: Update target allocations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newTargets]
              properties:
                newTargets:
                  type: array
                  items:
                    $ref: '#/components/schemas/TargetAllocation'
                  minItems: 2
      responses:
        '200':
          description: Targets updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/simulate:
    post:
      tags: [Portfolio]
      summary: Simulate rebalance (dry-run)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      currentDrift:
                        type: object
                      plannedSwaps:
                        type: array
                        items:
                          type: object
                      estimatedSwapCount:
                        type: integer
                      wouldRebalance:
                        type: boolean
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/rebalance:
    post:
      tags: [Portfolio]
      summary: Trigger manual rebalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rebalance result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Rebalance failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/drift:
    get:
      tags: [Portfolio]
      summary: Check portfolio drift from targets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Drift data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      maxDrift:
                        type: number
                      drifts:
                        type: object
                        additionalProperties:
                          type: number
                      needsRebalance:
                        type: boolean
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/history:
    get:
      tags: [Portfolio]
      summary: Get rebalance history for a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rebalance history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        swapLegs:
                          type: array
                          items:
                            type: object
                        driftBefore:
                          type: string
                        timestamp:
                          type: string
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/trigger-mode:
    put:
      tags: [Portfolio]
      summary: Update trigger mode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [triggerMode]
              properties:
                triggerMode:
                  $ref: '#/components/schemas/TriggerMode'
      responses:
        '200':
          description: Trigger mode updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/deactivate:
    post:
      tags: [Portfolio]
      summary: Deactivate a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Portfolio deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/activate:
    post:
      tags: [Portfolio]
      summary: Activate a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Portfolio activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Portfolio not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/portfolio/{id}/rebalance-detail:
    get:
      tags: [Portfolio]
      summary: Get rebalance detail by contract ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rebalance detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...